requirements.txt
torch>=2.0
torchvision
ultralytics
transformers
sentence-transformers
faiss-cpu
gradio>=4.0
moviepy
decord
opencv-python
numpy
Pillow
tqdm
einops=0.8.1 
timm=1.0.22
clip @ git+https://github.com/openai/CLIP.git